+++ src/actors/Projectile.cpp
@@ -1,4 +1,4 @@
 #include "Projectile.h"
 
-Projectile::Projectile(const std::filesystem::path &spriteLocation, double posX, double posY)
+Projectile::Projectile(const std::filesystem::path &spriteLocation, int posX, int posY)
     : Actor(spriteLocation, posX, posY) {}
